S7_new/pyproject.toml
[project]
name = "python-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.9.0",
    "llama-index>=0.12.28",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.1.0",
    "rich>=14.0.0",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "numpy>=1.24.0"
]
